refactor: 게임 시작, 비교, 재시작 메소드 클래스 Game 분리
